apply miss chance, critical attack, and armor defense

+++ Assets/Scripts/Stats/PlayerStats.cs
@@ -12,7 +12,7 @@ public class PlayerStats : StatsController
         player = GetComponent<Player>();
     }
 
-    public override void TakeDamage(float damage)
+    public override void TakeDamage(int damage)
     {
         base.TakeDamage(damage);
         player.DamageEffect();
